+++ src/components/linksViewer/style.scss
@@ -7,9 +7,11 @@
 
   .text-primary {
     font-size: $font-l;
+    color: $color-text-1;
   }
   
   .text-secondary {
     font-size: $font-m;
+    color: $color-text-3;
   }
 }
